*MAJOR* architecture overhaul:  - previously l10n_server and l10n_community were divided with the thinking    that a different editing interface might be used on the same data  - but real life discussions showed that a cool interface is supposed    to be tied to different data sources
So this upgrade separated the data sources (connectors) from the community interface.
Now only the previously implemented l10n_drupalorg connector is here, but it is easy
to implement others, as I'll do later.

There are also some smaller bugs fixed, but otherwise a huge reorganization was enough
for starters. I could not stop myself from cleaning up the suggestions architecture
at the same time though. So now the history and suggestions are not stored separately,
but are accessible from the same table (as inactive records for history values and as
suggestion records for suggestions). Otherwise they use the same data as translations,
so there was no reason to separate that out.

I also fixed smaller import and export issues, so these operations work better.

HTML and form structure cleanup is on the TODO list for the string editing interface,
which will be hopefully followed by a great UI improvement session later.